requirements.txt
requests
python-dotenv
idna
rapidfuzz